Update README.md
This is the very first project I did on NLP. The age detector processes the text written by a user to predict the writer's age. 
The model consists of a combination of a simple Neural Network (NN) and Linear Regression.  The Neural network has 50 hidden layers. The age predicted by this age detector has 8.98 mean absolute error, which means the model is quite accurate.
The data used in this project is from the following research: J. Schler, M. Koppel, S. Argamon and J. Pennebaker (2006). Effects of Age and Gender on Blogging in Proceedings of 2006 AAAI Spring Symposium on Computational Approaches for Analyzing Weblogs. 
The data can be downloaded from the link below:
https://www.kaggle.com/rtatman/blog-authorship-corpus